requirements.txt
# Django Book Management System - Requirements
# PostgreSQL Version - Production Ready

# Core Django
Django>=4.1.0,<5.0.0

# PostgreSQL Database Driver (REQUIRED)
psycopg2-binary>=2.9.5

# Image Processing (REQUIRED for ImageField)
Pillow>=10.0.0

# Environment Configuration (REQUIRED for .env support)
python-dotenv>=1.0.0

# Additional utilities (optional)
django-extensions>=3.2.0

# Development tools (optional)
ipython>=8.0.0

# Data migration tools
PyMySQL>=1.0.2

# Excel export functionality
openpyxl>=3.1.0

# Installation instructions:
# 1. Activate your conda environment: conda activate book_project
# 2. Remove MySQL packages: pip uninstall mysqlclient PyMySQL -y
# 3. Install core packages: conda install django pillow -c conda-forge -y
# 4. Install PostgreSQL driver: pip install psycopg2-binary
# 5. Install additional packages: pip install python-decouple django-extensions ipython PyMySQL

# PostgreSQL Setup Commands:
# 1. Install PostgreSQL Server 15+
# 2. Create database: CREATE DATABASE db_book;
# 3. Create user: CREATE USER bookuser WITH PASSWORD 'BookProject123!';
# 4. Grant privileges: GRANT ALL PRIVILEGES ON DATABASE db_book TO bookuser;
# 5. Use Django's dumpdata/loaddata for migration: python data_migration.py
